pyproject.toml
[tool.poetry]
name = "spotify-covers"
version = "0.1.0"
description = ""
authors = ["gordonjb <gordonjb@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^7.1.2"
tomlkit = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"

[tool.poetry.scripts]
covers = "spotify_covers.covers:main"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

